Remove scary and incorrect statement from commentary
> While this seemed like a good idea at first it ultimately led me
> declare .emacs bankruptcy and switch to just Custom.  Your milage
> might differ though, that's why I am making this available even
> though I don't plan to develop it any further.

I did not declare .emacs bankruptcy *because* of `elc', but *despite*.
I am using `use-package' now and for me the main benefit is that it
provides *less* flexibility.  Before I switched to `use-package' and
especially while using a literate programming approach, I was always
where uncertain how to structure my init file.  For example, if some
minor-mode should be enabled in several major-mode, do I add a new
section for that minor mode, or do I enable it in each of the relevant
major-mode sections.